Move gamepad button declarations into lib
