+++ packages/example/src/Wall.tsx
@@ -4,35 +4,28 @@ import {
   gamepadButton,
   keycode,
   KEYS,
+  GAMEPADS,
   mouseAxis,
   useAxis,
   useButtonHeld,
   useButtonPressed,
 } from 'use-control'
-import { useGamepadButtonPressed } from './gamepad'
-
-const XBOX_ONE = {
-  D_LEFT: 14,
-  D_RIGHT: 15,
-  D_UP: 12,
-  D_DOWN: 13,
-  STICK_L_X: 0,
-  STICK_L_Y: 1,
-  STICK_R_X: 2,
-  STICK_R_Y: 3,
-}
 
 const inputMap = {
   buttons: {
-    left: [keycode(KEYS.left_arrow), keycode(KEYS.a), gamepadButton(0, XBOX_ONE.D_LEFT)],
-    right: [keycode(KEYS.right_arrow), keycode(KEYS.d), gamepadButton(0, XBOX_ONE.D_RIGHT)],
-    up: [keycode(KEYS.up_arrow), keycode(KEYS.w), gamepadButton(0, XBOX_ONE.D_UP)],
-    down: [keycode(KEYS.down_arrow), keycode(KEYS.s), gamepadButton(0, XBOX_ONE.D_DOWN)],
+    left: [keycode(KEYS.left_arrow), keycode(KEYS.a), gamepadButton(0, GAMEPADS.XBOX_ONE.D_LEFT)],
+    right: [
+      keycode(KEYS.right_arrow),
+      keycode(KEYS.d),
+      gamepadButton(0, GAMEPADS.XBOX_ONE.D_RIGHT),
+    ],
+    up: [keycode(KEYS.up_arrow), keycode(KEYS.w), gamepadButton(0, GAMEPADS.XBOX_ONE.D_UP)],
+    down: [keycode(KEYS.down_arrow), keycode(KEYS.s), gamepadButton(0, GAMEPADS.XBOX_ONE.D_DOWN)],
     count: [keycode(KEYS.space)],
   },
   axes: {
-    x: [mouseAxis('x'), gamepadAxis(0, XBOX_ONE.STICK_R_X)],
-    y: [mouseAxis('y'), gamepadAxis(0, XBOX_ONE.STICK_R_Y)],
+    x: [mouseAxis('x'), gamepadAxis(0, GAMEPADS.XBOX_ONE.STICK_R_X)],
+    y: [mouseAxis('y'), gamepadAxis(0, GAMEPADS.XBOX_ONE.STICK_R_Y)],
   },
 }
 
@@ -44,12 +37,6 @@ function modify<T>(val: MutableRefObject<T> | undefined, sink: (v: T) => void) {
 
 export default function Wall(props: any) {
   const mesh = useRef()
-
-  useGamepadButtonPressed(0, XBOX_ONE.D_LEFT, () => {
-    modify<any>(mesh, (m) => {
-      m.scale.x -= 0.1
-    })
-  })
 
   useAxis(inputMap, 'x', (v) => {
     modify<any>(mesh, (m) => {
